Update to latest Rubocop

+++ app/controllers/sufia/single_use_links_viewer_controller.rb
@@ -9,11 +9,12 @@ module Sufia
 
       def initialize(user, single_use_link)
         @user = user || ::User.new
+        return unless single_use_link
 
         @single_use_link = single_use_link
         can :read, [ActiveFedora::Base, ::SolrDocument] do |obj|
           single_use_link.valid? && single_use_link.itemId == obj.id && single_use_link.destroy!
-        end if single_use_link
+        end
       end
     end
   end
